docker_learn_1/requirement.txt
weaviate-client==4.8.1
sentence-transformers==3.1.1
fastapi==0.115.0
uvicorn==0.30.6
huggingface-hub==0.25.1
